add function that read csv file and add medicines
